backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI & ML
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3

# Databases
pymongo==4.6.1
neo4j==5.15.0
motor==3.3.2

# APIs & HTTP
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4